tests: tests added and enabled for wasm
