+++ src/SamplesApp/UITests.Shared/Windows_UI_Xaml_Media/Geometry/ArcSegmentPage.xaml
@@ -1,5 +1,5 @@
 ﻿<Page
-    x:Class="SamplesApp.Windows_UI_Xaml_Media.Geometries.ArcSegmentPage"
+    x:Class="SamplesApp.Windows_UI_Xaml_Media.Geometry.ArcSegmentPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -9,7 +9,7 @@
 
 	<StackPanel>
 		<TextBlock Text="This a ArcSegment" />
-		<Path Stroke="Black" StrokeThickness="1">
+		<Path x:Name="Path" Stroke="Black" StrokeThickness="1">
 			<Path.Data>
 				<PathGeometry>
 					<PathGeometry.Figures>
