+++ src/SamplesApp/UITests.Shared/Windows_UI_Xaml_Media/Geometry/LineSegmentPage.xaml
@@ -1,5 +1,5 @@
 ﻿<Page
-    x:Class="SamplesApp.Windows_UI_Xaml_Media.Geometries.BezierSegmentPage"
+    x:Class="SamplesApp.Windows_UI_Xaml_Media.Geometry.LineSegmentPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -7,17 +7,25 @@
     mc:Ignorable="d"
     Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
 
-	<StackPanel>
-		<TextBlock Text="This a BezierSegment" />
-		<Path Stroke="Black" StrokeThickness="1">
+    <StackPanel>
+		<Path x:Name="Path" Stroke="Black" StrokeThickness="1">
 			<Path.Data>
 				<PathGeometry>
 					<PathGeometry.Figures>
 						<PathFigureCollection>
-							<PathFigure StartPoint="10,100">
+							<PathFigure IsClosed="True" StartPoint="10,100">
 								<PathFigure.Segments>
 									<PathSegmentCollection>
-										<BezierSegment Point1="100,0" Point2="200,200" Point3="300,100"/>
+										<LineSegment Point="100,100" />
+										<LineSegment Point="100,50" />
+									</PathSegmentCollection>
+								</PathFigure.Segments>
+							</PathFigure>
+							<PathFigure IsClosed="True" StartPoint="10,10">
+								<PathFigure.Segments>
+									<PathSegmentCollection>
+										<LineSegment Point="100,10" />
+										<LineSegment Point="100,40" />
 									</PathSegmentCollection>
 								</PathFigure.Segments>
 							</PathFigure>
